requirements.txt
colorama==0.4.6
playsound==1.3.0
pycryptodome==3.20.0
